requirements.txt
requests==2.28.1
python-dotenv==0.21.0
apscheduler==3.9.1
tweepy==4.10.1
python_graphql_client==0.4.3
Pillow==9.5.0
boto3==1.24.89
# jinja2==3.1.2